move definition of ImageLoaderManagement class to image-loader
