+++ image-loader/app/ImageLoaderComponents.scala
@@ -1,16 +1,15 @@
 import com.gu.mediaservice.GridClient
-import com.gu.mediaservice.lib.aws.{DynamoDB, SimpleSqsMessageConsumer}
-import com.gu.mediaservice.lib.config.{ServiceHosts, Services}
+import com.gu.mediaservice.lib.aws.SimpleSqsMessageConsumer
+import com.gu.mediaservice.lib.config.Services
 import com.gu.mediaservice.lib.imaging.ImageOperations
 import com.gu.mediaservice.lib.logging.GridLogging
-import com.gu.mediaservice.lib.management.{InnerServiceStatusCheckController, ImageLoaderManagement}
+import com.gu.mediaservice.lib.management.InnerServiceStatusCheckController
 import com.gu.mediaservice.lib.play.GridComponents
-import controllers.{ImageLoaderController, IngestQueueMonitorController, UploadStatusController}
+import controllers.{ImageLoaderController, ImageLoaderManagement, IngestQueueMonitorController, UploadStatusController}
 import lib._
 import lib.storage.{ImageLoaderStore, QuarantineStore}
 import model.{Projector, QuarantineUploader, Uploader}
 import play.api.ApplicationLoader.Context
-import play.api.libs.ws.WSClient
 import router.Routes
 
 class ImageLoaderComponents(context: Context) extends GridComponents(context, new ImageLoaderConfig(_)) with GridLogging {
